bug 1399639: Refactor AttachmentTemplatesTests
Refactor to pytest style, including moving some pytest fixtures into the
global conftest.py.
