+++ kuma/attachments/tests/test_templates.py
@@ -1,64 +1,47 @@
 import pytest
-from constance import config
 from pyquery import PyQuery as pq
 
 from kuma.core.urlresolvers import reverse
-from kuma.users.tests import UserTestCase
-from kuma.wiki.tests import document, revision, WikiTestCase
+from kuma.wiki.models import Revision
 
 from ..models import Attachment
 from . import make_test_file
 
 
-class AttachmentTemplatesTests(UserTestCase, WikiTestCase):
+@pytest.mark.security
+def test_xss_file_attachment_title(admin_client, constance_config, root_doc,
+                                   wiki_user, editor_client):
+    constance_config.WIKI_ATTACHMENT_ALLOWED_TYPES = 'text/plain'
 
-    def setUp(self):
-        super(AttachmentTemplatesTests, self).setUp()
-        self.old_allowed_types = config.WIKI_ATTACHMENT_ALLOWED_TYPES
-        config.WIKI_ATTACHMENT_ALLOWED_TYPES = 'text/plain'
-        self.client.login(username='admin', password='testpass')
-        self.document = document(save=True)
-        self.files_url = reverse('attachments.edit_attachment',
-                                 kwargs={'document_path': self.document.slug},
-                                 locale='en-US')
+    # use view to create new attachment
+    file_for_upload = make_test_file()
+    files_url = reverse('attachments.edit_attachment',
+                        kwargs={'document_path': root_doc.slug},
+                        locale='en-US')
+    title = '"><img src=x onerror=prompt(navigator.userAgent);>'
+    post_data = {
+        'title': title,
+        'description': 'xss',
+        'comment': 'xss',
+        'file': file_for_upload,
+    }
+    response = admin_client.post(files_url, data=post_data)
+    assert response.status_code == 302
 
-    def tearDown(self):
-        super(AttachmentTemplatesTests, self).tearDown()
-        config.WIKI_ATTACHMENT_ALLOWED_TYPES = self.old_allowed_types
+    # now stick it in/on a document
+    attachment = Attachment.objects.get(title=title)
+    content = '<img src="%s" />' % attachment.get_file_url()
+    root_doc.current_revision = Revision.objects.create(
+        document=root_doc, creator=wiki_user, content=content)
 
-    @pytest.mark.security
-    def test_xss_file_attachment_title(self):
-        title = '"><img src=x onerror=prompt(navigator.userAgent);>'
-        # use view to create new attachment
-        file_for_upload = make_test_file()
-        post_data = {
-            'title': title,
-            'description': 'xss',
-            'comment': 'xss',
-            'file': file_for_upload,
-        }
-        self.client.login(username='admin', password='testpass')
-        response = self.client.post(self.files_url, data=post_data)
-        self.assertEqual(response.status_code, 302)
-
-        # now stick it in/on a document
-        attachment = Attachment.objects.get(title=title)
-        rev = revision(content='<img src="%s" />' % attachment.get_file_url(),
-                       document=self.document,
-                       save=True)
-
-        # view it and verify markup is escaped
-        response = self.client.get(rev.document.get_edit_url())
-        self.assertEqual(response.status_code, 200)
-        doc = pq(response.content)
-        self.assertEqual(
-            '%s xss' % title,
-            doc('.page-attachments-table .attachment-name-cell').text()
-        )
-        self.assertIn(
-            '&gt;&lt;img src=x onerror=prompt(navigator.userAgent);&gt;',
-            doc('.page-attachments-table .attachment-name-cell').html()
-        )
-        # security bug 1272791
-        for script in doc('script'):
-            self.assertNotIn(title, script.text_content())
+    # view it and verify markup is escaped
+    response = editor_client.get(root_doc.get_edit_url())
+    assert response.status_code == 200
+    doc = pq(response.content)
+    text = doc('.page-attachments-table .attachment-name-cell').text()
+    assert text == ('%s xss' % title)
+    html = doc('.page-attachments-table .attachment-name-cell').html()
+    assert '&gt;&lt;img src=x onerror=prompt(navigator.userAgent);&gt;' in html
+    # security bug 1272791
+    for script in doc('script'):
+        assert title not in script.text_content()
