+++ kuma/conftest.py
@@ -3,6 +3,7 @@ from datetime import datetime
 import pytest
 from django.conf import settings
 from django.core.cache import caches
+from waffle.models import Flag
 
 from kuma.wiki.models import Document, Revision
 
@@ -12,6 +13,33 @@ def cleared_cacheback_cache():
     caches[settings.CACHEBACK_CACHE_ALIAS].clear()
     yield
     caches[settings.CACHEBACK_CACHE_ALIAS].clear()
+
+
+class ConstanceConfigWrapper(object):
+    """A Constance configuration wrapper to allow overriding the config."""
+    _original_values = []
+
+    def __setattr__(self, attr, value):
+        from constance import config
+        self._original_values.append((attr, getattr(config, attr)))
+        setattr(config, attr, value)
+        # This can fail if Constance uses a cached database backend
+        # CONSTANCE_DATABASE_CACHE_BACKEND = False to disable
+        assert getattr(config, attr) == value
+
+    def finalize(self):
+        from constance import config
+        for attr, value in reversed(self._original_values):
+            setattr(config, attr, value)
+        del self._original_values[:]
+
+
+@pytest.fixture
+def constance_config(db, settings):
+    """A Constance config object which restores changes after the testrun."""
+    wrapper = ConstanceConfigWrapper()
+    yield wrapper
+    wrapper.finalize()
 
 
 @pytest.fixture
@@ -47,6 +75,12 @@ def user_client(client, wiki_user):
     wiki_user.save()
     client.login(username=wiki_user.username, password='password')
     return client
+
+
+@pytest.fixture
+def editor_client(user_client):
+    Flag.objects.create(name='kumaediting', everyone=True)
+    return user_client
 
 
 @pytest.fixture
